+++ Runtime/JsonHelpers/CategoryConverter.cs
@@ -4,11 +4,11 @@ using Newtonsoft.Json.Linq;
 
 namespace ReadyPlayerMe.AvatarCreator
 {
-    public class AssetTypeConverter : JsonConverter
+    public class CategoryConverter : JsonConverter
     {
         public override bool CanConvert(Type objectType)
         {
-            return objectType == typeof(AssetType);
+            return objectType == typeof(Category);
         }
 
         public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
@@ -19,11 +19,17 @@ namespace ReadyPlayerMe.AvatarCreator
         public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
         {
             var token = JToken.Load(reader);
-            if (token.Type == JTokenType.String)
+            if (token.Type != JTokenType.String)
             {
-                return AssetTypeHelper.PartnerAssetTypeMap[token.ToString()];
+                throw new JsonSerializationException("Expected string value");
             }
-            throw new JsonSerializationException("Expected string value");
+            
+            if (!CategoryHelper.PartnerCategoryMap.ContainsKey(token.ToString()))
+            {
+                return Category.None;
+            }
+                
+            return CategoryHelper.PartnerCategoryMap[token.ToString()];
         }
     }
 }