+++ Runtime/JsonHelpers/PartnerAssetsDictionaryConverter.cs
@@ -5,9 +5,9 @@ using Newtonsoft.Json.Linq;
 
 namespace ReadyPlayerMe.AvatarCreator
 {
-    public class AssetTypeDictionaryConverter : JsonConverter<Dictionary<AssetType, object>>
+    public class CategoryDictionaryConverter : JsonConverter<Dictionary<Category, object>>
     {
-        public override void WriteJson(JsonWriter writer, Dictionary<AssetType, object> value, JsonSerializer serializer)
+        public override void WriteJson(JsonWriter writer, Dictionary<Category, object> value, JsonSerializer serializer)
         {
             var newValue = new Dictionary<string, object>();
             foreach (var element in value)
@@ -20,11 +20,11 @@ namespace ReadyPlayerMe.AvatarCreator
             serializer.Serialize(writer, newValue);
         }
 
-        public override Dictionary<AssetType, object> ReadJson(JsonReader reader, Type objectType,
-            Dictionary<AssetType, object> existingValue, bool hasExistingValue, JsonSerializer serializer)
+        public override Dictionary<Category, object> ReadJson(JsonReader reader, Type objectType,
+            Dictionary<Category, object> existingValue, bool hasExistingValue, JsonSerializer serializer)
         {
             var token = JToken.Load(reader);
-            var assets = new Dictionary<AssetType, object>();
+            var assets = new Dictionary<Category, object>();
             if (token.Type == JTokenType.Object)
             {
                 foreach (var element in token.ToObject<Dictionary<string, object>>())
@@ -35,12 +35,12 @@ namespace ReadyPlayerMe.AvatarCreator
                     }
 
                     var pascalCaseKey = char.ToUpperInvariant(element.Key[0]) + element.Key.Substring(1);
-                    if (!Enum.IsDefined(typeof(AssetType), pascalCaseKey))
+                    if (!Enum.IsDefined(typeof(Category), pascalCaseKey))
                     {
                         continue;
                     }
-                    var assetType = (AssetType) Enum.Parse(typeof(AssetType), pascalCaseKey);
-                    assets.Add(assetType, element.Value);
+                    var category = (Category) Enum.Parse(typeof(Category), pascalCaseKey);
+                    assets.Add(category, element.Value);
                 }
             }