+++ Runtime/PartnerAssetsManager.cs
@@ -13,53 +13,70 @@ namespace ReadyPlayerMe.AvatarCreator
     /// </summary>
     public class PartnerAssetsManager : IDisposable
     {
-        private const string EYE_MASK_SIZE_PARAM = "?w=256";
-
-        private readonly BodyType bodyType;
-        private readonly OutfitGender gender;
+        private const string TAG = nameof(PartnerAssetsManager);
+        private const string EYE_MASK_SIZE_SIZE = "?w=256";
+        private const string ASSET_ICON_SIZE = "?w=64";
 
         private readonly PartnerAssetsRequests partnerAssetsRequests;
-        private readonly CancellationTokenSource ctxSource;
-
+        private readonly Dictionary<Category, List<PartnerAsset>> assetsByCategory;
         public Action<string> OnError { get; set; }
 
-        private PartnerAsset[] assets;
+        private BodyType bodyType;
+        private OutfitGender gender;
+        private CancellationTokenSource ctxSource;
 
-        public PartnerAssetsManager(string partner, BodyType bodyType, OutfitGender gender, CancellationToken token = default)
+        public PartnerAssetsManager()
         {
-            this.bodyType = bodyType;
-            this.gender = gender;
+            partnerAssetsRequests = new PartnerAssetsRequests(CoreSettingsHandler.CoreSettings.AppId);
+            assetsByCategory = new Dictionary<Category, List<PartnerAsset>>();
+        }
+
+        public void SetAvatarProperties(BodyType assetBodyType, OutfitGender assetGender, CancellationToken token = default)
+        {
+            bodyType = assetBodyType;
+            gender = assetGender;
             ctxSource = CancellationTokenSource.CreateLinkedTokenSource(token);
-            partnerAssetsRequests = new PartnerAssetsRequests(partner);
         }
 
-        public async Task<Dictionary<string, AssetType>> GetAllAssets()
+        public async Task GetAssets()
         {
-            try
+            var startTime = Time.time;
+
+            var assets = await partnerAssetsRequests.Get(bodyType, gender, ctxSource.Token);
+            foreach (PartnerAsset asset in assets)
             {
-                assets = await partnerAssetsRequests.Get(ctxSource.Token);
-            }
-            catch (Exception e)
-            {
-                OnError?.Invoke(e.Message);
-                return null;
+                if (assetsByCategory.TryGetValue(asset.Category, out List<PartnerAsset> value))
+                {
+                    value.Add(asset);
+                }
+                else
+                {
+                    assetsByCategory.Add(asset.Category, new List<PartnerAsset> { asset });
+                }
             }
 
-            assets = assets.Where(FilterAssets).ToArray();
-            return assets.ToDictionary(asset => asset.Id, asset => asset.AssetType);
+            if (assets.Length != 0)
+            {
+                SDKLogger.Log(TAG, $"All asset received: {Time.time - startTime:F2}s");
+            }
         }
 
-        public async void DownloadAssetsIcon(Action<Dictionary<string, Texture>> onDownload)
+        public List<string> GetAssetsByCategory(Category category)
         {
-            var ordererAssets = assets.OrderByDescending(x => x.AssetType == AssetType.FaceShape);
-            var chunkList = ordererAssets.ChunkBy(20);
+            return assetsByCategory.TryGetValue(category, out List<PartnerAsset> _) ? assetsByCategory[category].Select(x => x.Id).ToList() : new List<string>();
+        }
 
+        public async Task DownloadAssetsIcon(Category category, Action<string, Texture> onDownload)
+        {
+            var startTime = Time.time;
+            var chunkList = assetsByCategory[category].ChunkBy(20);
+            
             foreach (var list in chunkList)
             {
-                Dictionary<string, Texture> assetIdTextureMap;
                 try
                 {
-                    assetIdTextureMap = await DownloadIcons(list);
+                    await DownloadIcons(list, onDownload);
+                    SDKLogger.Log(TAG, $"Download chunk of {category} icons: " + (Time.time - startTime) + "s");
                 }
                 catch (Exception e)
                 {
@@ -71,27 +88,34 @@ namespace ReadyPlayerMe.AvatarCreator
                 {
                     return;
                 }
-
-                onDownload?.Invoke(assetIdTextureMap);
                 await Task.Yield();
             }
         }
 
-        public bool IsLockedAssetCategories(string id)
+        public bool IsLockedAssetCategories(Category category, string id)
         {
-            var asset = assets.FirstOrDefault(x => x.Id == id);
+            if (!assetsByCategory.ContainsKey(category))
+            {
+                return false;
+            }
+
+            var asset = assetsByCategory[category].FirstOrDefault(x => x.Id == id);
             return asset.LockedCategories != null && asset.LockedCategories.Length > 0;
         }
 
-        private async Task<Dictionary<string, Texture>> DownloadIcons(List<PartnerAsset> chunk)
+        private async Task DownloadIcons(List<PartnerAsset> chunk, Action<string, Texture> onDownload)
         {
             var assetIconMap = new Dictionary<string, Task<Texture>>();
 
             foreach (var asset in chunk)
             {
-                var url = asset.AssetType == AssetType.EyeColor ? asset.Mask + EYE_MASK_SIZE_PARAM : asset.Icon;
+                var url = asset.Category == Category.EyeColor ? asset.Mask + EYE_MASK_SIZE_SIZE : asset.Icon + ASSET_ICON_SIZE;
                 var linkedTokenSource = CancellationTokenSource.CreateLinkedTokenSource(ctxSource.Token);
-                var iconTask = partnerAssetsRequests.GetAssetIcon(url, linkedTokenSource.Token);
+                var iconTask = partnerAssetsRequests.GetAssetIcon(url, icon =>
+                    {
+                        onDownload?.Invoke(asset.Id, icon);
+                    },
+                    linkedTokenSource.Token);
                 assetIconMap.Add(asset.Id, iconTask);
             }
 
@@ -99,31 +123,14 @@ namespace ReadyPlayerMe.AvatarCreator
             {
                 await Task.Yield();
             }
-
-            return assetIconMap.ToDictionary(a => a.Key, a => a.Value.Result);
         }
 
-        private bool FilterAssets(PartnerAsset asset)
+        public void DeleteAssets()
         {
-            // Outfit is only for full body and Shirt is only for half body.
-            // Both outfit and shirt are gender specific.
-            if (bodyType == BodyType.HalfBody)
-            {
-                if (asset.AssetType == AssetType.Shirt)
-                    return asset.Gender == gender;
-
-                return asset.AssetType != AssetType.Outfit;
-            }
-
-            if (asset.AssetType == AssetType.Outfit)
-                return asset.Gender == gender;
-
-            return asset.AssetType != AssetType.Shirt;
-        }
-
-        public void Dispose()
-        {
+            assetsByCategory.Clear();
             ctxSource?.Cancel();
         }
+
+        public void Dispose() => DeleteAssets();
     }
 }