+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+
+
+## [1.0.0] - 2023 August 31
+
+### Added
+- Analytics events
+- New asset endpoint
+
+### Updated
+- Asset loading strategy
+- Renamed asset type to category
+
 ## [0.5.0] - 2023 August 11
 
 ### Updated